refactor(sidecar): the two flat modules that were never products
composite.py is scene reading, not a product: native resolution per band, RGB
and false-colour composition, the continuous ramps. Six consumers across three
subjects reached it -- terra/actions.py, terra/energy/overlays.py, water.py and
three test modules -- which is what a shared reader looks like from the outside.
It joins sentinel2 and indices under terra/imagery/, and terra/energy/overlays.py
stops importing a flat module.

train_prithvi.py trains the Prithvi head. No action reaches it, it is not run
from the bundle, and it imported terra.actions and terra.imagery.sentinel2 from
outside the package. It moves to sidecar/tools/ beside the two generators, and
takes the same sys.path insert they use.

The attribute guard now covers tools/. Those scripts read across the package
boundary and nothing else checks that the names they read still exist: with
train_prithvi.py under tools/ the previous exclusion would have stopped
watching four of them.

Verification: pytest 513 passed, 8 skipped. ruff check clean, both import
contracts kept, gofmt and go vet clean. A ping through the entry point answers.
