+++ sidecar/terra/imagery/composite.py
